+++ modules/Users/Save.php
@@ -369,6 +369,9 @@ if(!$current_user->is_admin && !$GLOBALS['current_user']->isAdminForModule('User
 			$focus->setPreference('default_email_charset', $_REQUEST['default_email_charset'], 0, 'global');
 
 		if(isset($_POST['calendar_publish_key'])) $focus->setPreference('calendar_publish_key',$_POST['calendar_publish_key'], 0, 'global');
+                if(isset($_POST['subtheme'])) {
+                    $focus->setPreference('subtheme', $_POST['subtheme'], 0, 'global');
+                }
 	}
 
 	if (!$focus->verify_data())
